refactor(query): colocate expression coordination tests
