+++ crates/tinymist-query/src/analysis/signature.rs
@@ -413,87 +413,3 @@ fn analyze_dyn_signature(
 
     Some(func_signature(func))
 }
-
-#[cfg(test)]
-mod signature_tests {
-
-    use core::fmt;
-
-    use typst::syntax::LinkedNode;
-    use typst_shim::syntax::LinkedNodeExt;
-
-    use crate::analysis::{Signature, SignatureTarget, analyze_signature};
-    use crate::syntax::classify_syntax;
-    use crate::tests::*;
-
-    #[test]
-    fn test() {
-        snapshot_testing_at("..", "signature", &|ctx, path| {
-            let source = ctx.source_by_path(&path).unwrap();
-
-            let pos = ctx
-                .to_typst_pos(find_test_position(&source), &source)
-                .unwrap();
-
-            let root = LinkedNode::new(source.root());
-            let callee_node = root.leaf_at_compat(pos).unwrap();
-            let callee_node = classify_syntax(callee_node, pos).unwrap();
-            let callee_node = callee_node.node();
-
-            let result = analyze_signature(
-                ctx.shared(),
-                SignatureTarget::Syntax(source.clone(), callee_node.span()),
-            );
-
-            assert_snapshot!(SignatureSnapshot(result.as_ref()));
-        });
-    }
-
-    struct SignatureSnapshot<'a>(pub Option<&'a Signature>);
-
-    impl fmt::Display for SignatureSnapshot<'_> {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            let Some(sig) = self.0 else {
-                return write!(f, "<nil>");
-            };
-
-            let primary_sig = match sig {
-                Signature::Primary(sig) => sig,
-                Signature::Partial(sig) => {
-                    for w in &sig.with_stack {
-                        write!(f, "with ")?;
-                        for arg in &w.items {
-                            if let Some(name) = &arg.name {
-                                write!(f, "{name}: ")?;
-                            }
-                            let term = arg.term.as_ref();
-                            let term = term.and_then(|v| v.describe()).unwrap_or_default();
-                            write!(f, "{term}, ")?;
-                        }
-                        f.write_str("\n")?;
-                    }
-
-                    &sig.signature
-                }
-            };
-
-            writeln!(f, "fn(")?;
-            for param in primary_sig.pos() {
-                writeln!(f, " {},", param.name)?;
-            }
-            for param in primary_sig.named() {
-                if let Some(expr) = &param.default {
-                    writeln!(f, " {}: {},", param.name, expr)?;
-                } else {
-                    writeln!(f, " {},", param.name)?;
-                }
-            }
-            if let Some(param) = primary_sig.rest() {
-                writeln!(f, " ...{}, ", param.name)?;
-            }
-            write!(f, ")")?;
-
-            Ok(())
-        }
-    }
-}
