+++ crates/tinymist-query/src/diagnostics.rs
@@ -271,27 +271,3 @@ impl DiagnosticRefiner for OutOfRootHintRefiner {
         raw.with_hint("Cannot read file outside of project root.")
     }
 }
-
-#[cfg(test)]
-mod lint_tests {
-    use std::collections::BTreeMap;
-
-    use tinymist_lint::KnownIssues;
-
-    use crate::tests::*;
-
-    #[test]
-    fn test() {
-        snapshot_testing_at(".", "lint", &|ctx, path| {
-            let source = ctx.source_by_path(&path).unwrap();
-
-            let result = ctx.lint(&source, &KnownIssues::default());
-            let result = crate::diagnostics::DiagWorker::new(ctx).convert_all(result.iter());
-            let result = result
-                .into_iter()
-                .map(|(k, v)| (file_uri_(&k), v))
-                .collect::<BTreeMap<_, _>>();
-            assert_snapshot!(JsonRepr::new_redacted(result, &REDACT_LOC));
-        });
-    }
-}
