+++ crates/tinymist-query/src/syntax/module.rs
@@ -148,57 +148,3 @@ pub(crate) fn scan_workspace_files<T>(
 
     res
 }
-
-#[cfg(test)]
-mod module_tests {
-    use serde_json::json;
-    use tinymist_std::path::unix_slash;
-    use typst::syntax::FileId;
-
-    use crate::syntax::module::*;
-    use crate::tests::*;
-
-    #[test]
-    fn test() {
-        snapshot_testing_at("..", "modules", &|ctx, _| {
-            fn ids(ids: EcoVec<FileId>) -> Vec<String> {
-                let mut ids: Vec<String> = ids
-                    .into_iter()
-                    .map(|id| unix_slash(id.vpath().as_rooted_path_compat()))
-                    .collect();
-                ids.sort();
-                ids
-            }
-
-            let dependencies = construct_module_dependencies(ctx);
-
-            let mut dependencies = dependencies
-                .into_iter()
-                .map(|(id, v)| {
-                    (
-                        unix_slash(id.vpath().as_rooted_path_compat()),
-                        ids(v.dependencies),
-                        ids(v.dependents),
-                    )
-                })
-                .collect::<Vec<_>>();
-
-            dependencies.sort();
-            // remove /main.typ
-            dependencies.retain(|(path, _, _)| path != "/main.typ");
-
-            let dependencies = dependencies
-                .into_iter()
-                .map(|(id, deps, dependents)| {
-                    let mut mp = serde_json::Map::new();
-                    mp.insert("id".to_string(), json!(id));
-                    mp.insert("dependencies".to_string(), json!(deps));
-                    mp.insert("dependents".to_string(), json!(dependents));
-                    json!(mp)
-                })
-                .collect::<Vec<_>>();
-
-            assert_snapshot!(JsonRepr::new_pure(dependencies));
-        });
-    }
-}
